pyproject.toml
[project]
name = "test-api"
version = "0.1.0"
description = ""
authors = [
    {name = "BimBimBamBam1337",email = "a94440003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-jose (==3.3.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "skuf (>=0.3.0,<0.4.0)"
]

[tool.poetry]
packages = [{include = "test_api", from = "src"}]

[tool.alembic]
script_location = "%(here)s/src/infrastructure/database/migration"
prepend_sys_path = [
    "."
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
